moving FlowSolverFactory to separate file
