+++ src/testing/fd_state_sensitivity_wrt_parameter.h
@@ -15,7 +15,7 @@
 #include "parameters/all_parameters.h"
 #include "dg/dg_factory.hpp"
 #include "ode_solver/ode_solver_factory.h"
-#include "flow_solver/flow_solver.h"
+#include "flow_solver/flow_solver_factory.h"
 #include <fstream>
 
 namespace PHiLiP {
